requirements.txt
# Core ML and NLP libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
evaluate>=0.4.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Text processing
nltk>=3.8.0
langdetect>=1.0.9
simplemma>=0.9.0
sentencepiece>=0.1.99

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web scraping and API
requests>=2.31.0
beautifulsoup4>=4.12.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=6.5.0

# Optional: GPU support (uncomment if using CUDA)
# torch-audio>=2.0.0
# torchvision>=0.15.0
